renamed qrcode-gen file for QOL
